requirements.txt
# WEB / API
django
django-filter
djangorestframework
djangorestframework-simplejwt
django-cors-headers
drf-spectacular

# Database
django-tenants
psycopg[binary]

# Environment
python-dotenv

# Image handling
Pillow

# Google OAuth Dependencies
authlib
google-auth
google-auth-httplib2
google-auth-oauthlib
